% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_tree.R
\name{construct_tree}
\alias{construct_tree}
\title{Construct Tree}
\usage{
construct_tree(p, high_dim = ifelse(p > 6, T, F), ntree = 100)
}
\arguments{
\item{p}{Number of variables.}

\item{high_dim}{Logical value indicating whether the tree should be constructed in high dimension or not.}

\item{ntree}{Number of trees to be constructed.}
}
\value{
A list of edges representing the constructed trees.
}
\description{
Constructs a tree based on the given parameters.
}
\details{
This function constructs a tree based on the specified number of variables (\code{p}),
whether it should be constructed in high dimension (\code{high_dim}), and the number of trees to be constructed (\code{ntree}).
The function returns a list of edges representing the constructed trees.
}
\examples{
construct_tree(10, TRUE, 100)

construct_tree(5)

}
\seealso{
\code{\link{combn}}, \code{\link{plyr::alply}}, \code{\link{igraph::graph_from_edgelist}},
\code{\link{igraph::has_eulerian_cycle}}
}
