% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efi.R
\name{efi}
\alias{efi}
\title{Ensemble Fractional Imputation}
\usage{
efi(Y, dp, freq = F)
}
\arguments{
\item{Y}{categorical data with missingnes.Missing valuesa are encoded as NA. If Freq = T, the last column serves as the integer frequencies of the grouped observations.}

\item{dp}{a class doublep computed from \code{doublep}}

\item{freq}{an optional logical value that indicates whether frequency column exists in the data frame Y.}
}
\value{
A class  \code{"efi"}
}
\description{
Ensemble Fractional Imputation
}
\examples{
data(HairEyeColor)
p = 3
Y = do.call("rbind", apply(as.data.frame.table(HairEyeColor), 1, function(x) matrix(rep(x[1:p], each = x[p+1]), nc = p)))
Y = as.data.frame(Y)
for(k in 1:p){
Y[[k]] = factor(Y[[k]])
}
names(Y) <- names(dimnames(HairEyeColor))
(n = nrow(Y)); sum(HairEyeColor)
delta = matrix(rbinom(n * p, 1, 0.5), nr = n, nc = p)
Y[delta == 0] = NA

cand.edges = as.list(data.frame(combn(p, 2)))
dp = doublep(Y, cand.edges, freq = F)
EFI = efi(Y, dp, freq = F)
}
