% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.doublep}
\alias{plot.doublep}
\title{plot class doubledp}
\usage{
\method{plot}{doublep}(dp)
}
\arguments{
\item{dp}{a class doublep computed from \code{doublep}}
}
\description{
plot class doubledp
}
\examples{
data(HairEyeColor)
p = 3
Y = do.call("rbind", apply(as.data.frame.table(HairEyeColor), 1, function(x) matrix(rep(x[1:p], each = x[p+1]), nc = p)))
Y = as.data.frame(Y)
for(k in 1:p){
Y[[k]] = factor(Y[[k]])
}
names(Y) <- names(dimnames(HairEyeColor))
# (n = nrow(Y)); sum(HairEyeColor)
delta = matrix(rbinom(n * p, 1, 0.5), nr = n, nc = p)
Y[delta == 0] = NA

cand.edges = as.list(data.frame(combn(p, 2)))
dp = doublep(Y, cand.edges, freq = F)
plot(dp)
}
