% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate the joint probabilities}
\usage{
estimate(efi.obj, expr)
}
\arguments{
\item{efi.obj}{An object produced by \code{efi}.}

\item{expr}{A string that contains formulas with the desired probabilities.}
}
\value{
Estimates and Standard error of the parameters.
}
\description{
Estimate the joint probabilities
}
\examples{
data(HairEyeColor)
p = 3
Y = do.call("rbind", apply(as.data.frame.table(HairEyeColor), 1, function(x) matrix(rep(x[1:p], each = x[p+1]), nc = p)))
Y = as.data.frame(Y)
for(k in 1:p){
Y[[k]] = factor(Y[[k]])
}
names(Y) <- names(dimnames(HairEyeColor))
(n = nrow(Y)); sum(HairEyeColor)
delta = matrix(rbinom(n * p, 1, 0.5), nr = n, nc = p)
Y[delta == 0] = NA

cand.edges = as.list(data.frame(combn(p, 2)))
dp = doublep(Y, cand.edges, freq = F)
EFI = efi(Y, dp, freq = F)
estimate(EFI, "(Hair == \"Black\") & (Eye == \"Brown\")")
estimate(EFI, "(Hair == \"Black\") & (Sex == \"Male\")")
}
